pyproject.toml
[project]
name = "agent-google"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.1"
dependencies = [
    "google-adk>=1.22.1",
    "google-genai[agents,gui]>=1.59.0",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
